splash page logic setup
