+++ elaichi/lib/auth/domain/repository/auth_repository.dart
@@ -1,4 +1,5 @@
 import 'package:dartz/dartz.dart';
+import 'package:elaichi/auth/domain/datamodel/user_model.dart';
 import 'package:elaichi/auth/domain/repository/auth_failure.dart';
 import 'package:firebase_auth/firebase_auth.dart';
 import 'package:google_sign_in/google_sign_in.dart';
@@ -30,8 +31,8 @@ class AuthenticationRepository {
         accessToken: googleAuthentication.accessToken,
         idToken: googleAuthentication.idToken,
       );
-
       await _firebaseAuth.signInWithCredential(authCrendential);
+      UserData.instance().user = _firebaseAuth.currentUser;
     } on FirebaseException catch (e) {
       throw LogInWithGoogleFailure.fromCode(e.code);
     } catch (_) {
