+++ elaichi/pubspec.yaml
@@ -27,6 +27,8 @@ dependencies:
   font_awesome_flutter: ^9.2.0
   lottie: ^1.2.1
 
+
+
   # Firebase Plugins
   google_sign_in: ^5.2.1
   firebase_auth: ^3.2.0
